Extract SelectCardView from SelectCardFragment.
